Adding rudimentary attack system
Weapon attacks are stored in a data asset containing an array of "WeaponMoves" each weapon move is associated with an animation and has the array index of potential combo followups stored. Anim notifies are used to drive different attack states such as damaging, not damaging, and ready for combo. These must be fed back into the character using public functions by the anim blueprint