Quick fix to combo status desynchronization
Because the attack ability end was replicated, the client would replicate the end of the ability and prevent the server from resetting it's version of the combo index. (The ability would have already ended so the server would never reach the animation end)

My quick fix for this is simply not replicating the end of the ability caused on animation end, and instead letting both the server and client predict the end of the ability.